Moved testsutils inside utils dir
